Changed the width and height of sprites to be float rather than int

+++ src/sprite.c
@@ -1,7 +1,7 @@
 #include "sprite.h"
 
 Sprite sprite_init(jmp_buf error, vec3 pos, unsigned int scale, char *texture_location,
-		vec2 translation[], vec2 spr_num[], int instance_amount, int width, int height) {
+		vec2 translation[], vec2 spr_num[], int instance_amount, float width, float height) {
 
 	Sprite sprite;
 
@@ -29,7 +29,7 @@ void sprite_delete(Sprite *sprite) {
 }
 
 
-void sprite_generate_model(Model *model, int width, int height) {
+void sprite_generate_model(Model *model, float width, float height) {
 	// vertex, uv, normal
 	float vertices[8 * 4] = {
 		0, 0, 0,		0, 0,	0, 0, 0,
